+++ types/std_types.h
@@ -54,3 +54,14 @@ namespace fresa
 #else
     #error "ranges not supported"
 #endif
+
+//* coroutines
+#if __has_include(<coroutine>)
+    #include <coroutine>
+    namespace std_ = std;
+#elif __has_include(<experimental/coroutine>)
+    #include <experimental/coroutine>
+    namespace std_ = std::experimental;
+#else
+    #error "coroutines not supported"
+#endif