+++ CHANGELOG.md
@@ -11,8 +11,9 @@ _general description_
 - **added** - string utilities (lowercase, split)
 - **changed** - tests are now in standalone files in the tests/ directory and can be run using arguments "-t test1,test2"
 - **added** - coroutine implementation (future, promise, awaitables)
-- **in progress** - job system
+- **added** - job system
 - **added** - temporary dependency `jthread` while compilers gain supports
+- **added** - thread-safe atomic queue
 
 #### [0.4.0] core features (_16 jun 22_)